+++ backend/src/budgets/budgets.models.rs
@@ -4,7 +4,7 @@ use chrono::{DateTime, Local};
 use serde::{Deserialize, Serialize};
 use sqlx::{self, FromRow};
 
-#[derive(Serialize, FromRow, Debug)]
+#[derive(Serialize, FromRow)]
 pub struct Budget {
     pub budget_id: i64,
     pub user_id: i64,
@@ -12,7 +12,8 @@ pub struct Budget {
     pub year: i64,
 }
 
-#[derive(Serialize, FromRow, Debug)]
+#[derive(Serialize, FromRow)]
+#[serde(rename_all = "camelCase")]
 pub struct BudgetRowData {
     pub budget_id: i64,
     pub user_id: i64,
@@ -34,6 +35,7 @@ pub struct BudgetRowData {
 }
 
 #[derive(Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
 pub struct TransactionData {
     pub transaction_id: i64,
     pub title: String,
@@ -44,6 +46,7 @@ pub struct TransactionData {
 }
 
 #[derive(Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
 pub struct LineItemData {
     pub line_item_id: i64,
     pub name: String,
@@ -53,7 +56,6 @@ pub struct LineItemData {
     pub transactions: Vec<TransactionData>,
 }
 
-#[derive(Serialize, Deserialize, Clone)]
 pub struct BudgetCategoryDataMap {
     pub budget_category_id: i64,
     pub name: String,
@@ -61,13 +63,15 @@ pub struct BudgetCategoryDataMap {
 }
 
 #[derive(Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
 pub struct BudgetCategoryDataConverted {
     pub budget_category_id: i64,
     pub name: String,
-    pub budget_line_items: Vec<LineItemData>,
+    pub line_items: Vec<LineItemData>,
 }
 
-#[derive(Serialize, Deserialize, Clone)]
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
 pub struct BudgetResponseData {
     pub budget_id: i64,
     pub month_number: i64,
