+++ backend/src/budgets/budgets.service.rs
@@ -1,13 +1,8 @@
-use std::collections::HashMap;
-
 use actix_web::web::Data;
 
 use crate::AppState;
 
-use super::budgets_models::{
-    BudgetCategoryDataConverted, BudgetCategoryDataMap, BudgetRowData, LineItemData,
-    TransactionData,
-};
+use super::budgets_models::BudgetRowData;
 
 pub async fn get_budget(
     state: Data<AppState>,
@@ -34,13 +29,14 @@ pub async fn get_budget(
             tr.amount,
             tr.notes,
             tr.date
-        FROM budgets b
+        FROM 
+            budgets b
         JOIN 
             budget_categories bc ON b.id = bc.budget_id
         JOIN 
             line_items li ON bc.id = li.budget_category_id
         JOIN
-        	transactions tr ON li.id = tr.line_item_id 
+            transactions tr ON li.id = tr.line_item_id 
         WHERE 
             b.user_id = $1 AND b.month_number = $2 AND b.year = $3
     ";
@@ -51,48 +47,4 @@ pub async fn get_budget(
         .bind(year)
         .fetch_all(&state.db)
         .await
-}
-
-pub fn get_compiled_budget_data(rows: Vec<BudgetRowData>) -> Vec<BudgetCategoryDataConverted> {
-    let mut budget_categories_compiled: HashMap<i64, BudgetCategoryDataMap> = HashMap::new();
-
-    for row in rows {
-        let budget_category = budget_categories_compiled
-            .entry(row.budget_category_id)
-            .or_insert_with(|| BudgetCategoryDataMap {
-                budget_category_id: row.budget_category_id,
-                name: row.budget_category_name,
-                budget_line_items: HashMap::new(),
-            });
-
-        let line_item = budget_category
-            .budget_line_items
-            .entry(row.line_item_id)
-            .or_insert_with(|| LineItemData {
-                line_item_id: row.line_item_id,
-                name: row.line_item_name,
-                is_fund: row.is_fund,
-                starting_balance: row.starting_balance,
-                planned_amount: row.planned_amount,
-                transactions: Vec::new(),
-            });
-
-        line_item.transactions.push(TransactionData {
-            transaction_id: row.transaction_id,
-            title: row.title,
-            merchant: row.merchant,
-            amount: row.amount,
-            notes: row.notes,
-            date: row.date,
-        });
-    }
-
-    return budget_categories_compiled
-        .into_values()
-        .map(|category| BudgetCategoryDataConverted {
-            budget_category_id: category.budget_category_id,
-            name: category.name,
-            budget_line_items: category.budget_line_items.into_values().collect(),
-        })
-        .collect();
 }
